Add additional transformState getters
This change adds a latLngToScreenCoordinate overload, to retrieve
the projected vec4 in additional to the ScreenCoordinate object,
that is useful to detect whether the projected latLng is in front
or behind the camera.
The change also adds an additional, optional argument to getProjMatrix,
to retrieve a projection matrix centered around 0,0, instead of around
the current map center. This is necessary to minimize numerical error
in floating point math when projecting float coordinates with a float
matrix4 over a large domain.